Move addServiceConnection tests to its own file
The test_services_graphql_api.py file was starting to grow too big.
