+++ services/tests/test_services_graphql_api.py
@@ -3,12 +3,7 @@ from string import Template
 import pytest
 from django.conf import settings
 
-from open_city_profile.consts import (
-    SERVICE_CONNECTION_ALREADY_EXISTS_ERROR,
-    SERVICE_NOT_IDENTIFIED_ERROR,
-)
 from open_city_profile.graphene import TranslationLanguage
-from open_city_profile.tests.asserts import assert_match_error_code
 from open_city_profile.tests.graphql_test_helpers import do_graphql_call_as_user
 from services.enums import ServiceType
 from services.tests.factories import ProfileFactory, ServiceConnectionFactory
@@ -131,113 +126,6 @@ def test_service_connections_of_service_always_returns_an_empty_result(
     }
     executed = user_gql_client.execute(query)
     assert executed["data"] == expected_data
-
-
-@pytest.mark.parametrize("service__service_type", [ServiceType.BERTH])
-def test_normal_user_can_add_service(user_gql_client, service):
-    ProfileFactory(user=user_gql_client.user)
-
-    # service object with type is included in query just to ensure that it has NO affect
-    query = """
-        mutation {
-            addServiceConnection(input: {
-                serviceConnection: {
-                    service: {
-                        type: GODCHILDREN_OF_CULTURE
-                    }
-                    enabled: false
-                }
-            }) {
-                serviceConnection {
-                    service {
-                        type
-                        name
-                    }
-                    enabled
-                }
-            }
-        }
-    """
-
-    expected_data = {
-        "addServiceConnection": {
-            "serviceConnection": {
-                "service": {"type": service.service_type.name, "name": service.name},
-                "enabled": False,
-            }
-        }
-    }
-    executed = user_gql_client.execute(query, service=service)
-    assert executed["data"] == expected_data
-
-
-@pytest.mark.parametrize("service__service_type", [ServiceType.BERTH])
-def test_normal_user_cannot_add_service_multiple_times_mutation(
-    user_gql_client, service
-):
-    ProfileFactory(user=user_gql_client.user)
-
-    query = """
-        mutation {
-            addServiceConnection(input: {
-                serviceConnection: {
-                }
-            }) {
-                serviceConnection {
-                    service {
-                        type
-                        name
-                    }
-                }
-            }
-        }
-    """
-
-    expected_data = {
-        "addServiceConnection": {
-            "serviceConnection": {
-                "service": {"type": service.service_type.name, "name": service.name}
-            }
-        }
-    }
-    executed = user_gql_client.execute(query, service=service)
-    assert dict(executed["data"]) == expected_data
-    assert "errors" not in executed
-
-    # do the mutation again
-    executed = user_gql_client.execute(query, service=service)
-    assert "errors" in executed
-    assert "code" in executed["errors"][0]["extensions"]
-    assert (
-        executed["errors"][0]["extensions"]["code"]
-        == SERVICE_CONNECTION_ALREADY_EXISTS_ERROR
-    )
-
-
-def test_not_identifying_service_for_add_service_connection_produces_service_not_identified_error(
-    user_gql_client,
-):
-    ProfileFactory(user=user_gql_client.user)
-
-    query = """
-        mutation {
-            addServiceConnection(input: {
-                serviceConnection: {
-                }
-            }) {
-                serviceConnection {
-                    service {
-                        type
-                        name
-                    }
-                }
-            }
-        }
-    """
-
-    executed = user_gql_client.execute(query, service=None)
-
-    assert_match_error_code(executed, SERVICE_NOT_IDENTIFIED_ERROR)
 
 
 def test_normal_user_can_query_own_services_gdpr_api_scopes(
